Upgrade stable branch to Rust 1.87.0 (#4065)
Advance the stable branch to 7a126c2b2e1 from main and use the 1.87.0
channel.

This PR was created through doing the following on the [stable
branch](https://github.com/model-checking/kani/tree/stable):
1. Running: `git merge 7a126c2b2e1  -X theirs`
2. Modifying the channel in `rust-toolchain.toml` to `1.87.0`
3. Running `cargo update -p deranged` and `cargo update -p tokio` to
resolve the failures in the deny workflow

The rationale for picking 7a126c2b2e1 is as follows:
- The 1.87.0 release was branched from main on 2025-03-28 (see
https://releases.rs/docs/1.87.0/)
- The latest commit on main for which compilation and regressions pass
with the 1.87.0 channel is 7a126c2b2e1 (which uses the 2025-04-02
toolchain).
- The next commit (1ff0c8f5626) upgrades the toolchain to 2025-04-03 and
updates to expected files. If we were to use that commit, we would need
to revert the updates to those files.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.