Add setup scripts for Ubuntu 20.04 (#4082)
Background:
1. Kani currently cannot be installed using `cargo install
kani-verifier` on Ubuntu 20.04 because of the older glibc version, and
thus needs to be built from source.
2. The `scripts/setup/ubuntu-20.04` directory is a symlink to
`scripts/setup/ubuntu` which is incorrect because the installation
scripts in the `ubuntu` directory assume the existence of a deb package
for CBMC which is not the case for Ubuntu 20.04.
 
To make it easier for users to build from source on Ubuntu 20.04, this
PR adds the necessary setup scripts under a new
`scripts/setup/ubuntu-20.04` directory.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.