Update toolchain to 2025-05-22 (#4098)
Culprit PR: https://github.com/rust-lang/rust/pull/139916, which removed
some of the debug assertions, and hence we have fewer property checks. I
just deleted the numbers entirely to make the tests less brittle; it
doesn't matter how many properties we check as long as we catch the UB.

Resolves #4093

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.