Update `kani::mem` pointer validity documentation (#4092)
https://github.com/rust-lang/rust/pull/117329/files#diff-c175d4e27676febf62c061d31cf9756d256b46e2e44cc6b3177d4ff75e932567L18-L29
changed the pointer safety documentation. We changed our models
accordingly:


https://github.com/model-checking/kani/blob/92bc45da85ff70edb5a341f5566f150721150247/library/kani_core/src/mem.rs#L186-L190

but forgot to update the documentation.

See https://doc.rust-lang.org/beta/core/ptr/index.html for the current
documentation; the point of this PR is to 1) update our docs to match
this reference and 2) state which of the points Kani can verify.

Resolves #4087

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.