Add support for edition 2018 crates using assert! (Fixes #3717) (#4096)
We export `core::assert` as `__kani__workaround_core_assert!` to every
crate which uses `assert`. Because kani uses edition 2021, crates which
are built edition 2018 now use the 2021 `assert`.

The behaviour of `assert` changed between both editions (see
https://doc.rust-lang.org/nightly/edition-guide/rust-2021/panic-macro-consistency.html).

This commit adds an extra clause to handle the case, if the first
message parameter isn't of string type and adds a format string
accordingly

This resolves #3717 . 

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.

---------

Co-authored-by: Zyad Hassan <zyadh@amazon.com>
Co-authored-by: Zyad Hassan <88045115+zhassan-aws@users.noreply.github.com>