Handle generic defaults in BoundedArbitrary derives (#4117)
For structs with defaults for the generic parameters, Kani's
`#[derive(BoundedArbitrary)]` macro was generating incorrect code. This
PR fixes the issue through ensuring that the defaults are not included
in the where clause or the impl definition.

Resolves #4116

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.