Improve linking error output for `#[no_std]` crates (#4126)
Adds a more descriptive error message when kani is run on a `no_std`
project that hasn't imported the `kani` crate.

Previously there would just be an internal assertion failure (or a
generic error when compiling in release mode), but the kani compiler now
detects this specific case and points to the simple solution of adding
`extern crate kani` as suggested
[here](https://github.com/model-checking/kani/issues/3906#issuecomment-2680374095).
This PR also introduces a regression test for this case to the
`cargo-ui` suite.

Resolves #3906 

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.

---------

Co-authored-by: Carolyn Zech <carolynzech@gmail.com>
Co-authored-by: Carolyn Zech <cmzech@amazon.com>