Fix the git log command in the toolchain update script (#4139)
The `git log` command used in the toolchain update script generated
commits that are not between the current and next toolchain commits.
This PR fixes the command through adding the `--ancestry-path` option:
```
       --ancestry-path
           When given a range of commits to display (e.g.  commit1..commit2 or commit2 ^commit1), only display commits that exist directly on the ancestry chain between the commit1 and commit2, i.e. commits that are both descendants of commit1, and ancestors of commit2.
```
Resolves #4128 

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.