Optimize goto binary exporting in `cprover_bindings` (#4148)
Motivated by the finding that the
[`write_goto_binary_file`](https://github.com/model-checking/kani/blob/1b25a1853fab3262380e9baefdfeb929edd53fd1/cprover_bindings/src/irep/goto_binary_serde.rs#L97-L103)
function is responsible for around half of our compiler's codegen time,
this PR does a collection of low-hanging fruit optimizations for the way
we export goto files in the `cprover_bindings` crate.

This includes:
- Switching `IrepNumbering` caches to use the faster
[`FxHashMap`](https://nnethercote.github.io/2021/12/08/a-brutally-effective-hash-function-in-rust.html)
(as used in the rust compiler)
- Using `.entry().or_insert()` to initialize the irep cache
- This avoids having to hash separately for a `.get()` and `.insert()`
(as the compiler seemed to have not optimized them together)
- Making `IrepId` strings copy-on-write
- This avoids unnecessary string allocations by allowing us to construct
`InternedString`s directly from borrowed string literals when possible

Initial testing (on
[`perf/format`](https://github.com/model-checking/kani/blob/main/tests/perf/format/src/lib.rs)
&
[`perf/vec/vec`](https://github.com/model-checking/kani/blob/main/tests/perf/vec/vec/src/main.rs))
suggests that these changes make our compiler ~13% faster, but I'm also
working on comparing performance on the whole std project for the
future.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.