Add the option to generate performance flamegraphs (#4138)
Adds the ability for `cargo kani` to optionally output the info needed
for generating flamegraphs of the compiler & driver's performance. When
enabled with the `FLAMEGRAPH` environment variable, it will generate an
output file for the driver and one for each crate in the workspace as
they are compiled. These output files can then be rendered as
flamegraphs to see where Kani is spending most of its execution time.
See [this draft docs
page](https://github.com/AlexanderPortland/kani/blob/flamegraph/docs/src/profiling.md)
for more detailed instructions on how it would be used.

This should help us pinpoint which pieces of the compilation process are
most to blame for our compilation times and focus future efforts to
improve them.

Resolves #4075

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.

---------

Co-authored-by: Carolyn Zech <carolynzech@gmail.com>
Co-authored-by: Carolyn Zech <cmzech@amazon.com>