Ensure that contract closures are FnOnce (#4151)
Rust believes that Kani's contract closures are `FnMut`. This prevents
us from writing contracts for functions that return mutable references
to their input arguments (#3764).

To ensure Rust correctly infers these closures as `FnOnce`, they need to
be wrapped in a dummy function that explicitly requires an `FnOnce`.
This wrapping must be done at the point of closure definition, as doing
it later, when calling the function, doesn't seem to have any effect.

Resolves #3764

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.

---------

Co-authored-by: Fedor Ryabinin <rfedor@amazon.com>
Co-authored-by: Carolyn Zech <cmzech@amazon.com>