Upgrade Rust toolchain to 2025-07-04 (#4199)
Relevant upstream PRs:
- https://github.com/rust-lang/rust/pull/143214 (Remove let_chains
unstable feature)
- https://github.com/rust-lang/rust/pull/143036 (Remove support for
`dyn*` from the compiler)

The latter was tracked as an unsupported feature in Kani, whereby we can
resolve one of our own issues as well.

Resolves: #4196
Resolves: #1784

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.