Add unstable option prove-safety-only (#4239)
Compute verification results under the assumption that no panic occurs.
This enables safety verification, i.e., proving the absence of undefined
behavior without being distracted by panics which Kani, otherwise, also
reports as verification failures. #4230 was an efforts towards the same
goal, but required manual annotation of the conditions under which a
panic may occur (and may hide undefined behavior when either the given
specifications are too broad or when undefined behavior arises on a path
to a panic).

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.