Introduce thread pool for writing goto binaries in parallel (#4236)
Serializing and writing goto binaries is a serious bottleneck for the
Kani compiler, with profiling indicating that it takes around half of
the total codegen time.

This PR introduces a variable size thread pool specifically for
serializing goto binaries and writing them to disk. Now, instead of
having to do everything itself, the main compiler thread just has
collect all the data needed for serialization (requiring relatively
inexpensive clones of some local state) and dispatch it to the thread
pool's work queue. This allows the main thread to move on quickly while
the pool's worker threads handle generating the binaries off of the
critical path of compilation.

### Results
The table below shows **wall clock end-to-end compile times** before and
after this change. This metric corresponds to how long a user would have
to wait for Kani's compilation to finish before verification can begin.
| benchmark    | compile time before | compile time now | change |
| ---  | --- | --- | --- |
| standard library (commit
[`177d0fd`](https://github.com/model-checking/verify-rust-std/commit/177d0fd25fbf3cb82c7b8408140158af1e408ede))
| 328s | 233s | -95s (-29%) |
| `prost` crate (from #2505) | 325s | 177s | -148s (-45%) |

_from local runs on a 12 core M3 Mac_

Resolves #2505.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.