Add loop invariant support for `while let` loop  (#4279)
This PR adds loop invariant support for `while let` loop, and also fix
the bug reported in issue #4275

I don't separate them because the bug fix is simple and necessary for
the expected test of `while let` loop.

Resolves #4275
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.