[Breaking Change] Fail if stub verified doesn't have a contract harness (#4295)
Emit a compilation error if the target of a `stub_verified` attribute
does not have a contract harness. Also require `-Z stubbing` for
stub_verified. Both of these are breaking changes. The former I feel
strongly should go in, since IMO it's a soundness hole. The latter I
could be convinced to revert if we feel like it's unnecessary churn; it
just felt odd to me to have an unstable stub feature that's not behind
the unstable stub flag.

Note that this compilation error is not really sufficient to ensure that
the stub's contract holds, since a user can just pass `--harness` to
skip the contract harness. A better design would be to insert the check
automatically rather than requiring a separate harness. But this is
better than nothing and less invasive to implement, so start with this
for now.

Resolves https://github.com/model-checking/kani/issues/4294

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.