Fix LLBC regressions (#4338)
Since the upgrade to the 2025-07-04 toolchain, which removed the
`let_chains` unstable feature, the LLBC regressions have been broken.
This was caused by the removal of the `let_chains` feature (which is
used in the Charon submodule) and requiring the 2024 edition to use let
chains.

The edition used in Charon was upgraded to 2024 since
https://github.com/AeneasVerif/charon/commit/59b8ad61cdf1ad02bd16f9af5bdff696b095c66f,
but the attempt to update Kani to use that commit ran into
https://github.com/AeneasVerif/charon/issues/806. So, in the meantime,
apply a minimal patch to the Charon code after cloning it to allow it to
build.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.