Combo of small performance changes (#4314)
This PR is a pair of small performance changes. It used to be 4 (and
thus would've been a real combo), but two involved caching so I've moved
them under #4313 instead. These changes include:
1. passing information commonly used by our `GotocHook`s' `hook_applies`
method as arguments, so that each hook doesn't have to recompute it on
it's own
2. a small one-line change to compute a fn instance's name once instead
of twice.

These are both very small, but together with the other two changes now
moved to #4313, made a ~6% decrease in the compile time of the standard
library.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.

---------

Co-authored-by: Felipe R. Monteiro <felisous@amazon.com>