Revert "Cache dependencies for CI jobs (#4181)" (#4375)
This reverts commit ada8ed3d9b3e22f7005ed18746fe8372644c2122: caching
.rustup means we just keep piling up toolchains. We already have 6 GB as
cache size, and will eventually just hit the 10 GB limit and not be
caching anything anymore.

We could revisit caching to build something more effective, but also
possible effects of caching are somewhat negligible anyway.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.