fix: Make kani attribute nameres work with generic args having `::` (#4427)
I found this while working on
https://github.com/model-checking/verify-rust-std/issues/382.

When you try to add `#[proof_for_contract(Rc::<MaybeUninit<T>,
A>::assume_init)]` to a proof for this function
https://doc.rust-lang.org/1.90.0/src/alloc/rc.rs.html#1211, kani fails
to resolve `Rc::<MaybeUninit<T>, A>::assume_init`, as it compares the
*last two parts* of the path string,


https://github.com/model-checking/kani/blob/e0316ff45f7ccf724b70e28af80f719b784cfcc2/kani-compiler/src/kani_middle/resolve.rs#L783-L797

But rustc's `TyCtxt::def_path_str` emits the path
`rc::Rc::<core::mem::MaybeUninit<T>, A>::assmue_init` even if there's
`use core::mem::MaybeUninit` in scope. So the *last two parts* of it is
`::MaybeUninit<T>, A>::assmue_init`, instead of being
`::<core::mem::MaybeUninit<T>, A>::assmue_init`, and the resolution
fails.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.

Co-authored-by: Felipe R. Monteiro <felisous@amazon.com>