NixOS: patch binaries if the dynamic linker is a stub (#4413)
On modern NixOS installations `/lib64/ld-linux-x86-64.so.2` seems to be
a link to `/nix/store/%hash%-stub-ld-x86_64-unknown-linux-musl`, which
prints the following when invoked:
> Could not start dynamically linked executable: [...]
> NixOS cannot run dynamically linked executables intended for generic
> linux environments out of the box. For more information, see:
> https://nix.dev/permalink/stub-ld

But the current check assumes that things are fine and does not patch
anything.

This PR makes `kani` still patch binaries if
`/lib64/ld-linux-x86-64.so.2` is a stub.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.

Co-authored-by: Felipe R. Monteiro <felisous@amazon.com>
Co-authored-by: Michael Tautschnig <tautschn@amazon.com>