docs: Correct `default-unwind` Cargo.toml examples (#4496)
The examples for configuring `default-unwind` inside Cargo.toml files
use unquoted integers. This isn't how they are actually expected to be
written, and following these examples leads to kani failing to run.

This corrects the examples in the documentation to align with the
expected configuration format.

Resolves #4495

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.

Signed-off-by: hashcatHitman <155700084+hashcatHitman@users.noreply.github.com>